requirements.txt
git+https://github.com/saleweaver/python-sp-api.git
python-dotenv
pymysql
pandas
psycopg2-binary